services/favorites_service/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
ecdsa==0.19.1
fastapi==0.115.13
greenlet==3.2.3
grpcio==1.73.0
grpcio-tools==1.73.0
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
